requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
PyPDF2==3.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
cryptography==43.0.0


